Update logging.md
I spent a lot of time trying to get the trace working, and it was because the application pool did not have write access to the program directory by default. There were no errors thrown to indicate this, just no trace file was written. This may be obvious to others, but it certainly wasn't for me.