More detail on how to hook up the InMemoryUserList
I had an issue where I couldn't figure out how to hook up the InMemoryUserList to IdentityServer and I kept reading over line 323 ("In Startup replace the empty user list with a call to the Get method.") without realizing that line 323 was the step that I was missing.  I think this code (or maybe even a smaller snippet of my code) should be added in to prevent people from accidentally skipping over the step.